Add changes to successful work
For the correct work and operation of the application, I changed the JavaConfig
configuration and instead of using db and hibernate.properties, I
started using static final variables. In addition, I edited the html
page's for correct display. And in general, I changed the rest of the
files so that they correctly and correctly performed their role