Added changes to the mentor's comments
After all, I added the db.properties file according to the mentor's
comments and finally my code began to work correctly. I also fixed the
jambs (by type: when I updated the user, instead of updating the current
user, I created a new one, but with changed data. And to that, I changed
the dependency to pom.xml and my Validator is working, now when
creating/changing a user, you cannot create an invalid user)